+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [x.x.x] -
 
 ### Added
--  Info endpoint (/) for API v1
+- Info endpoint (/) for API v1
 - Add new datasets using the command line
+- Update existing datasets using the command line